Melhorando as notificações  e incluindo confirmações de ação

+++ templates/admin/pages/view.tpl.php
@@ -1,19 +1,19 @@
 <h3 class="mb-5">Detalhes da página</h3>
 
 <div class="row">
-    <div class="col-3">
+    <div class="col-md-6">
         <dl class="row">
-            <dt class="col-sm-6">Título</dt>
-            <dd class="col-sm-10">Página inicial</dd>
+            <dt class="col-sm-4">Título</dt>
+            <dd class="col-sm-8">Página inicial</dd>
             
-            <dt class="col-sm-6">URL</dt>
-            <dd class="col-sm-10">/ - <a href="/" target="blank">abrir</a></dd>
+            <dt class="col-sm-4">URL</dt>
+            <dd class="col-sm-8">/ - <a href="/" target="blank">abrir</a></dd>
 
-            <dt class="col-sm-6">Criado em</dt>
-            <dd class="col-sm-10">2018-01-01 01:01</dd>
+            <dt class="col-sm-4">Criado em</dt>
+            <dd class="col-sm-8">2018-01-01 01:01</dd>
 
-            <dt class="col-sm-6">Atualizado em</dt>
-            <dd class="col-sm-10">2019-01-01</dd>
+            <dt class="col-sm-4">Atualizado em</dt>
+            <dd class="col-sm-8">2019-01-01</dd>
 
         </dl>
 
@@ -27,7 +27,7 @@
 
 <p>
     <a href="/admin/pages/1/edit" class="btn btn-primary">editar</a>
-    <a href="/admin/pages/1/delete" class="btn btn-danger">remover</a>
+    <a href="/admin/pages/1/delete" class="btn btn-danger confirm">remover</a>
 
 </p>
